requirements.txt
fastapi
uvicorn
asyncpg
psycopg2
sqlalchemy
alembic
fastapi-sqlalchemy
pydantic
python-dotenv
python-multipart
passlib
python-jose
bcrypt